refactor: boardType을 뷰 패키지로 이동
