+++ src/test/java/janggi/dao/dao/JdbcPieceDaoTest.java
@@ -6,7 +6,6 @@ import janggi.dao.DatabaseTest;
 import janggi.dao.piece.PieceEntity;
 import janggi.model.Team;
 import janggi.model.board.Board;
-import janggi.model.board.BoardType;
 import janggi.model.piece.Byeong;
 import janggi.model.piece.Piece;
 import janggi.model.piece.palace.Jang;
@@ -14,6 +13,7 @@ import janggi.model.piece.palace.Sa;
 import janggi.model.position.absolute.Column;
 import janggi.model.position.absolute.Position;
 import janggi.model.position.absolute.Row;
+import janggi.view.mapping.BoardType;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
